move to a lib so we can call from code not cli
